load pretraining dataset from huggingface, select text column
sourcing datasets for pretraining: we plan to combine two different datasets for our pretraining. One, a pretraining dataset from huggingface. Two, a custom dataset that we scrape from the web (in this project, we use randomly scraped python code from Github). At this commit, we load the pretraining dataset from huggingface and in the next step we get the scraped data.