define contrastive loss function
contrastive loss function helps in training models to understand similarity relationships between different inputs. here,  it ensures that questions are embedded close to their correct answers and far from incorrect ones. data: This is a matrix where each row represents the similarity scores between one question and all possible answers. For example, data[i, j] would represent the similarity score between the i-th question and the j-th answer. target = torch.arange(data.size(0)): This line creates the target tensor. For a batch of size N, the target tensor is [0, 1, 2, ..., N-1]. This implies that the i-th question should be most similar to the i-th answer.loss = torch.nn.CrossEntropyLoss()(data, target): The CrossEntropyLoss function is applied here. It treats each row of data as the output of a classification task, where the goal is to classify the row such that the highest score corresponds to the correct answer. The correct answer for the i-th row (question) is at index i.